video slicing done; args are added to save storage and time.

+++ slice_video.py
@@ -2,6 +2,7 @@ from depth2loc.utils.helper import *
 import cv2
 from tqdm import trange
 import subprocess
+import argparse
 
 def slice_video(rgb_ts_f_txt, segment_file):
     segments, acts = load_segment_file_to_datetime(segment_file) # by default, we have 38 actions
@@ -38,46 +39,47 @@ def slice_video(rgb_ts_f_txt, segment_file):
     return action_label_per_frame, actions_amount
 
 if __name__ == "__main__":
-    action_labels, act_amount = slice_video(rgb_ts_f_txt, seg_f_txt)
+    parser = argparse.ArgumentParser(description="Slice video into segments")
+    parser.add_argument("--slice_video", action="store_true", help="Whether to slice video into segments")
+    args = parser.parse_args()
+
+    action_labels_per_frame, act_amount = slice_video(rgb_ts_f_txt, seg_f_txt)
     if act_amount != action_amount:
         print(f"Warning: action amount {act_amount} != {action_amount} in params.yaml")
     else:
         print(f"Fortunately, we have {act_amount} actions as expected.")
 
-    ## using ffmpeg to accelerate video processing
 
-    rgb_parent_folder = rgb_f.rsplit('/', 1)[0]
+    if args.slice_video:
+        rgb_parent_folder = rgb_f.rsplit('/', 1)[0]
+        for i in trange(act_amount):
+            continued = True
+            has_start = False
+            has_end = False
+            os.remove(f"{rgb_parent_folder}/frames_{i:02d}.txt") if os.path.exists(f"{rgb_parent_folder}/frames_{i:02d}.txt") else None
+            os.remove(f"{rgb_parent_folder}/rgb_{i:02d}.mp4") if os.path.exists(f"{rgb_parent_folder}/rgb_{i:02d}.mp4") else None
+            for j in range(len(action_labels_per_frame)):
+                if action_labels_per_frame[j] == i:
+                    if not has_start:
+                        has_start = True
+                    if has_end:
+                        continued = False
+                    with open(f"{rgb_parent_folder}/frames_{i:02d}.txt", "a") as f:
+                        f.write(f"file '{rgb_F}/frame_{j:05d}.png'\n")
+                        f.write(f"duration {1/15}\n")
+                elif has_start and not has_end:
+                    has_end = True
 
-    ### remove all txt in <rgb_parent_folder>
-    
+            cmd = [
+                "ffmpeg",
+                "-f", "concat",
+                "-safe", "0",
+                "-i", f"{rgb_parent_folder}/frames_{i:02d}.txt",
+                "-r", "15",
+                "-c:v", "libx264",
+                "-pix_fmt", "yuv420p",
 
-    for i in trange(act_amount):
-        continued = True
-        has_start = False
-        has_end = False
-        os.remove(f"{rgb_parent_folder}/frames_{i:02d}.txt") if os.path.exists(f"{rgb_parent_folder}/frames_{i:02d}.txt") else None
-        os.remove(f"{rgb_parent_folder}/rgb_{i:02d}.mp4") if os.path.exists(f"{rgb_parent_folder}/rgb_{i:02d}.mp4") else None
-        for j in range(len(action_labels)):
-            if action_labels[j] == i:
-                if not has_start:
-                    has_start = True
-                if has_end:
-                    continued = False
-                with open(f"{rgb_parent_folder}/frames_{i:02d}.txt", "a") as f:
-                    f.write(f"file '{rgb_F}/frame_{j:05d}.png'\n")
-            elif has_start and not has_end:
-                has_end = True
-
-        cmd = [
-            "ffmpeg",
-            "-f", "concat",
-            "-safe", "0",
-            "-i", f"{rgb_parent_folder}/frames_{i:02d}.txt",
-            "-r", "15",
-            "-c:v", "libx264",
-            "-pix_fmt", "yuv420p",
-
-            f"{rgb_parent_folder}/rgb_{i:02d}.mp4"
-        ]
-        subprocess.run(cmd)
+                f"{rgb_parent_folder}/rgb_{i:02d}.mp4"
+            ]
+            subprocess.run(cmd)
